requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langgraph==0.0.15
langchain-google-genai==0.0.11
python-dotenv==1.0.1
gradio==3.41.0
# langchain==0.0.309